unci: use 32bit component_indices internally

+++ libheif/codecs/uncompressed/unc_boxes.h
@@ -109,7 +109,7 @@ public:
 
   struct Component
   {
-    uint16_t component_index;
+    uint32_t component_index;
     uint16_t component_bit_depth; // range [1..256]
     uint8_t component_format;
     uint8_t component_align_size;
